CPU: Refactor operation dispatch
Do away with unwieldy Operation enum and replace with a map within the
CPU. This allows use of wrapper functions to apply the flag handling
logic, and also makes it possible to create metafunctions for
constructing the various classes of operation (e.g. an
increment(Register) method for generating increment operations).