Implement Governor to control main loop waits
Track clock skew and tactically wait when it gets too large.

The Nintendo logo appears to scroll way too fast, but I think I suspect
the scrolling logic at this point - if I slow down the governor the logo
appears to jump between tiles in a way that the physical gameboy
doesn't.