Rework length counter API
The field previously called "remaining time" is better called "length
counter" - the value written to the NRx1 registers _initializes_ the
counter on the channel, but isn't actually used for the subsequent
countdown - it's up to the channel to store an in-memory copy and tick
it down.

As such the channels can't rely on querying Io for the current state of
the counter, and so this information needs to be passed via a listener.