Performance improvement to SpriteMap
Removing invalidated entries from rowToSpriteMap and patternToSpriteMap
was taking up a lot of time in the main loop as the current
implementation involved checking all values in the map to identify which
to remove.

New implementation maintains reverse maps, spriteToRowMap and
spriteToPatternMap, which are used solely to identify rows to clear
during invalidation.