Fix rounding errors in decimator
Previous implementation derived a modulus M by dividing the input and
output frequencies as integers, and then dropping M-1 in every M input
samples. This was susceptible to integer rounding error, leading to a
small mismatch between the generated output frequency and the frequency
expected by the renderer. This led to a build-up of samples in the
output buffer, leading to the latency between audio and gameplay
increasing over time without bound.

New implementation performs a division over doubles, and multiplies that
by the total number of ticks seen across all time, rounding the result
down to a long, R. Whenever the value of R differs from the previous R,
we emit that sample, but otherwise drop it.

This new implementation prevents calculation errors compounding over
time, and thus avoids the sample build-up.