Performance optimization to AbstractChannel
AbstractChannel is responsible, among other things, for uplifting mono
samples from the various channels into Stereo output. This means it
creates a lot of width-two arrays. Profiling hints that this array
creation might have been taking up a lot of time, so as an optimization
we now use a large lookup table to map samples and audio modes to the
corresponding (precalculated, unchanging) stereo output arrays.