Use audio stack to throttle main loop
Ditch previous approach of using Thread.sleep to adjust for clock skew.
Even on average, and regardless of the length of sleep requested,
Thread.sleep doesn't seem to reliably average out at the desired length
of sleep. On the two machines I've tried it completes reliably faster
than desired. This could have been a programming error, but if so I
couldn't find it.

New approach seems to give pretty much perfect timings, in exchange for
a small amount of visual chop if the audio stack chugs during scene
changes. This coud also probably be improved with tuning of values.