Crappy implementation of RAM saving
Persist extram to disk every 2 seconds because I was too tired to
implement proper caching/backoff and just wanted to get it working.

It works, but I definitely need to re-write bits of this, re-arrange
the code a bit, and delete the dead bits from when I used serialised
byte streams.